requirements.txt
# great-ideas/requirements.txt
streamlit
matplotlib
pillow
dspy-ai
python-dotenv
requests
pypi-json